Refactor: Apply Single Responsibility Principle to converter architecture
Replace monolithic SSLSessionSNIConverter with focused extractor classes
following SOLID design principles and best practices.

Changes:
- Create SNIHostnameExtractor: Dedicated SNI hostname extraction
- Create ClientCertificateCNExtractor: Dedicated client cert CN extraction
- Remove SSLSessionSNIConverter: Deprecated facade (backward compatibility)
- Remove SSLEngineClientCNConverter: Stale wrapper with no real purpose
- Update kafka-ssl-jfr.xml: Use new specialized converters directly
- Update integration tests: Use new extractor classes
- Update documentation: Reflect new architecture

Benefits:
- Single Responsibility: Each class has one clear purpose
- Better maintainability: Easier to modify and test independently
- Cleaner codebase: No deprecated code or thin wrappers
- Consistent design: All converters follow same pattern

All 28 tests passing (10 SNI, 12 ClientCN, 5 integration, 1 E2E)
